ENH: Always set segmentation nrrd file vector axis to "list" kind
When only a few segments were in the segmentation then axis kind was set to complex, RGB, RGBA, ..., which was not meaningful and caused some nrrd readers to fail to read the image (see this discussion for an example https://discourse.itk.org/t/reading-segmentation-file-from-slicer-nrrd-with-multiple-segments-with-simpleitk-shows-exception/808/14).

Solution: non-spatial axis kind is always set to "list".