BUG: avoid timer event backlog
Previously, a QTimer with a fixed interval was used
to trigger updates of the proxy nodes.  If the update
took longer than than the timeout interval, then it was
possible for timer events to backlog in the event
queue, making the application unresponsive to input
(specifically, it became impossible to push the pause
button to stop the timer).

With this change, the timer is used in singleShot mode, so that
a new timer request is made only after the proxy nodes have
been updated.  This fixes the backlog issue and should have no
noticable impact on playback speed if the proxy update is
brief.

STYLE: remove dead code and fix spacings

BUG: set otherwise unused variable to fix logic